@RackerWilliams 2016-10-24
Yup, I'm aware of the problem.  I'm currently working on a number of optimizations to address. I'm also planning to specify an "auto" for the xsd-engine config so that we always use Saxon when it's available.